Wild Bootstrap for Linear Model Coefficients

Performs wild bootstrap resampling for linear regression models to handle heteroscedasticity. Supports Rademacher and Mammen weight schemes.

Usage

wild_boot_lm(fit, R = 2000, type = c("rademacher", "mammen"))

Arguments

fit

an object of class 'lm' from lm. Should be fitted linear regression model. Heteroscedasticity is allowed and expected - this method is specifically designed to handle non-constant variance in residuals. Model should have at least 1 predictor.

R

integer number of bootstrap replicates (default 2000). Larger values (5000-10000) recommended for confidence intervals in publications. Must be >= 1 and a whole number.

type

character string specifying weight distribution scheme. Options: "rademacher" (default, faster, robust) generates weights as +1/-1 with equal probability. "mammen" (asymptotically optimal) uses empirically calibrated Mammen distribution with golden ratio. Choice has minimal practical impact on results.

Value

A list with two elements:

coef

numeric vector of original fitted model coefficients, including intercept if present. Names preserved from original model.

boot

numeric matrix of dimensions (p x R) where p is number of coefficients. Each column contains one bootstrap replicate of coefficients. Row names are coefficient names; column names are bootstrap iteration numbers.

Details

The wild bootstrap works by resampling residuals with random signs/weights while keeping predictors fixed. This is particularly useful for heteroscedastic data.

Examples

set.seed(42)
x <- rnorm(50)
y <- 2 + 1.5 * x + rnorm(50, sd = abs(x))  # heteroscedastic
fit <- lm(y ~ x)
result <- wild_boot_lm(fit, R = 500, type = "rademacher")
head(result$boot, 3)
#>           [,1]     [,2]     [,3]      [,4]      [,5]     [,6]     [,7]
#> [1,] 1.9753203 1.980576 2.123834 2.0678834 1.9076264 2.159163 2.210491
#> [2,] 0.8720766 1.067381 1.166961 0.6549819 0.9530609 1.113307 1.102528
#>           [,8]     [,9]     [,10]    [,11]    [,12]     [,13]     [,14]
#> [1,] 2.0071533 2.321284 2.0047517 1.817300 2.054271 1.9586098 1.8044803
#> [2,] 0.8469358 1.225791 0.6771568 0.983804 1.387659 0.8355168 0.9087932
#>          [,15]     [,16]    [,17]     [,18]     [,19]    [,20]    [,21]
#> [1,] 1.8772024 2.0914638 2.042053 1.8357163 2.0098004 2.180394 2.168897
#> [2,] 0.6343612 0.8315492 1.082773 0.6585681 0.8517919 1.121889 1.239114
#>          [,22]     [,23]     [,24]     [,25]    [,26]     [,27]     [,28]
#> [1,] 1.8678718 1.9645377 1.9704229 1.7682576 1.998697 2.0969814 1.9496607
#> [2,] 0.9088898 0.6524283 0.8549948 0.7656812 1.212092 0.8000851 0.9044014
#>          [,29]    [,30]     [,31]    [,32]    [,33]   [,34]    [,35]    [,36]
#> [1,] 1.9783842 2.192163 2.0267414 2.357393 2.138209 2.18538 2.195567 2.054828
#> [2,] 0.9838859 1.313850 0.9529679 1.137953 1.205997 1.20790 1.196713 1.161149
#>         [,37]     [,38]    [,39]    [,40]    [,41]     [,42]     [,43]
#> [1,] 2.117032 2.0435523 2.315509 2.013956 2.013385 1.8064654 1.8901772
#> [2,] 1.066317 0.9905804 1.040586 1.344549 1.020957 0.8439125 0.8651475
#>          [,44]    [,45]     [,46]    [,47]   [,48]     [,49]    [,50]     [,51]
#> [1,] 1.9283627 2.288276 2.1048421 2.092782 2.05462 1.9558288 2.228870 2.0212879
#> [2,] 0.8061694 1.388382 0.6969381 1.250057 1.24446 0.9548495 1.016961 0.5415847
#>         [,52]     [,53]    [,54]     [,55]     [,56]    [,57]    [,58]
#> [1,] 1.962542 2.0946488 2.237268 1.9062581 1.8008992 2.117271 2.119889
#> [2,] 1.107244 0.6255991 1.072414 0.5641755 0.8590504 1.177510 1.082598
#>          [,59]   [,60]    [,61]    [,62]    [,63]     [,64]    [,65]    [,66]
#> [1,] 1.8830591 2.15127 2.099798 1.684571 2.150763 1.8167414 2.038317 1.948357
#> [2,] 0.9210717 1.36471 1.222432 0.839589 1.043079 0.9207779 0.851484 1.149639
#>          [,67]    [,68]     [,69]     [,70]     [,71]     [,72]    [,73]
#> [1,] 2.1593290 2.105964 2.0943648 1.9313846 2.1241076 2.2451961 2.179193
#> [2,] 0.5840348 1.060346 0.7631625 0.8479635 0.9289189 0.9777076 1.169570
#>          [,74]     [,75]    [,76]    [,77]     [,78]    [,79]    [,80]    [,81]
#> [1,] 1.7643172 1.8771424 2.052777 2.124891 2.0546949 2.108868 1.986722 1.831585
#> [2,] 0.8559077 0.9410542 1.024007 1.041826 0.7214367 1.067159 1.238111 1.018284
#>          [,82]     [,83]     [,84]     [,85]     [,86]     [,87]    [,88]
#> [1,] 2.0962866 1.9352294 2.0437422 1.8365600 1.9043945 1.8233706 2.197257
#> [2,] 0.9034236 0.8285953 0.8633523 0.8502138 0.7247583 0.7716146 1.134524
#>         [,89]    [,90]    [,91]    [,92]     [,93]     [,94]     [,95]    [,96]
#> [1,] 2.141941 2.271543 1.921789 2.201452 1.9942677 1.9667654 1.8801358 1.967321
#> [2,] 1.315957 1.201866 0.811591 1.139949 0.6699445 0.8204902 0.7684598 0.813252
#>          [,97]    [,98]    [,99]    [,100]    [,101]   [,102]   [,103]   [,104]
#> [1,] 1.8627139 2.281291 1.992537 1.9125908 1.9169208 2.097481 2.168196 2.242369
#> [2,] 0.5311686 1.214547 1.154938 0.8747794 0.9228786 1.236093 1.341802 1.474567
#>        [,105]   [,106]   [,107]    [,108]    [,109]   [,110]    [,111]   [,112]
#> [1,] 1.973546 2.029023 2.310475 1.7991626 2.0525786 2.256141 1.8255313 2.238212
#> [2,] 1.296463 1.281217 1.166908 0.8557247 0.7344148 1.068255 0.7357524 1.264254
#>         [,113]   [,114]   [,115]   [,116]    [,117]    [,118]    [,119]
#> [1,] 2.2616591 2.062673 1.933255 1.966407 1.9913459 2.0861879 1.6334229
#> [2,] 0.8766757 1.378156 1.185373 1.172006 0.9869061 0.6766652 0.7072345
#>         [,120]    [,121]   [,122]    [,123]   [,124]   [,125]    [,126]
#> [1,] 2.0920805 1.9106769 2.179555 2.1535453 2.038902 2.136907 1.9913710
#> [2,] 0.7732496 0.9203223 1.045242 0.7965515 1.377409 1.191644 0.5790866
#>         [,127]    [,128]   [,129]    [,130]    [,131]   [,132]    [,133]
#> [1,] 1.8283486 2.0421527 2.204248 1.9120785 1.7047559 2.004249 2.1123271
#> [2,] 0.9587922 0.8783117 1.109595 0.7993865 0.6613539 1.166494 0.9641577
#>        [,134]    [,135]    [,136]    [,137]   [,138]    [,139]   [,140]
#> [1,] 2.062060 2.0432714 2.1572398 1.9221688 2.040981 1.9008063 1.867233
#> [2,] 1.089531 0.9446281 0.9090601 0.7094762 1.295614 0.9079429 1.112123
#>         [,141]    [,142]    [,143]   [,144]   [,145]   [,146]  [,147]    [,148]
#> [1,] 1.9807501 2.0757646 2.0710035 2.122437 2.260917 1.938297 2.07237 1.9473447
#> [2,] 0.9361025 0.9005154 0.6811756 1.077148 1.077796 1.060670 1.21733 0.9737523
#>        [,149]    [,150]   [,151]   [,152]   [,153]   [,154]    [,155]   [,156]
#> [1,] 1.682241 2.0335738 1.950805 2.022749 2.211587 1.750916 2.0685008 1.878408
#> [2,] 1.043553 0.8602739 1.081605 1.207716 1.145500 1.142140 0.8724625 1.043454
#>        [,157]   [,158]   [,159]   [,160]    [,161]   [,162]   [,163]    [,164]
#> [1,] 2.223997 2.270873 2.065595 2.000916 1.9353198 2.161811 2.089736 1.7333863
#> [2,] 1.312805 1.039959 1.266239 1.131241 0.8636346 1.124156 1.218492 0.9578672
#>         [,165]   [,166]   [,167]   [,168]   [,169]   [,170]   [,171]    [,172]
#> [1,] 1.8911552 2.195669 2.021325 1.898717 1.836768 1.942745 1.839864 1.7785464
#> [2,] 0.8766743 1.192263 1.268626 0.788938 1.133465 1.005398 1.014411 0.7282886
#>        [,173]   [,174]   [,175]   [,176]    [,177]   [,178]  [,179]    [,180]
#> [1,] 2.144661 2.146854 2.286519 2.041440 1.8856115 2.094999 2.23654 1.9503814
#> [2,] 1.129118 1.361982 1.121616 0.944405 0.9025294 1.249225 1.21956 0.5990817
#>        [,181]   [,182]    [,183]    [,184]   [,185]    [,186]    [,187]
#> [1,] 1.955411 2.003652 1.9081781 2.2135667 1.712501 1.8383105 2.0588093
#> [2,] 1.227784 0.873083 0.7645959 0.8879951 1.038236 0.8183481 0.8602543
#>         [,188]    [,189]    [,190]    [,191]   [,192]    [,193]   [,194]
#> [1,] 1.9442279 2.0060569 1.9150354 2.0182922 2.151957 1.9961951 2.151740
#> [2,] 0.7430061 0.7095198 0.9264295 0.9573519 1.414007 0.8828851 1.363828
#>         [,195]    [,196]   [,197]    [,198]    [,199]   [,200]   [,201]
#> [1,] 1.9513006 1.7466027 1.840858 1.8459872 1.8249970 2.172856 2.064302
#> [2,] 0.7887977 0.8343353 1.005735 0.8370263 0.8434235 1.144056 1.239362
#>        [,202]    [,203]    [,204]    [,205]   [,206]   [,207]    [,208]
#> [1,] 2.241219 2.1370881 1.7224610 1.9528165 2.254254 2.171355 1.8360984
#> [2,] 1.235364 0.7546055 0.8713406 0.8792706 1.157442 1.281156 0.7794522
#>         [,209]    [,210]   [,211]   [,212]    [,213]   [,214]    [,215]
#> [1,] 1.9632226 1.9179192 1.953582 2.071511 1.9912810 2.113516 1.9168354
#> [2,] 0.6597035 0.8143371 1.147270 1.032428 0.9412396 1.355064 0.6637986
#>         [,216]   [,217]   [,218]    [,219]   [,220]   [,221]   [,222]    [,223]
#> [1,] 1.8684754 2.095528 1.955670 2.0924464 2.113976 1.972859 2.024118 2.2274521
#> [2,] 0.9360161 1.264796 1.145641 0.8830382 1.027902 0.912486 1.326448 0.8774898
#>        [,224]   [,225]   [,226]    [,227]   [,228]   [,229]    [,230]    [,231]
#> [1,] 2.108816 2.007316 1.882239 1.9073613 2.160079 1.989483 1.7442859 2.1424648
#> [2,] 1.120827 1.134431 1.292324 0.7270975 1.469167 1.000681 0.8970007 0.9327985
#>         [,232]    [,233]    [,234]   [,235]    [,236]    [,237]    [,238]
#> [1,] 1.8803430 1.8844666 2.1951291 2.094768 1.8095067 1.9437828 1.8818799
#> [2,] 0.7525373 0.9665859 0.8863851 1.126717 0.7142375 0.6159868 0.9440129
#>         [,239]   [,240]   [,241]    [,242]    [,243]   [,244]   [,245]
#> [1,] 1.9783580 2.110063 1.762454 1.8892937 1.7660924 2.181421 2.127760
#> [2,] 0.9435218 1.375337 1.121812 0.7399275 0.7225649 1.070974 1.150919
#>         [,246]   [,247]   [,248]   [,249]   [,250]   [,251]   [,252]    [,253]
#> [1,] 1.8149813 2.222099 1.875441 2.172224 1.909724 1.734214 1.781153 1.7991748
#> [2,] 0.7996506 1.129435 1.165263 1.070419 1.029778 0.767603 0.770659 0.8911999
#>         [,254]    [,255]    [,256]    [,257]    [,258]   [,259]   [,260]
#> [1,] 1.7713794 2.2811886 2.1246378 2.0650837 1.9361347 2.110689 2.109484
#> [2,] 0.6479907 0.8935869 0.9417897 0.6753879 0.7868633 1.062097 1.110697
#>         [,261]    [,262]    [,263]   [,264]   [,265]    [,266]   [,267]
#> [1,] 1.8943345 2.0004303 1.7827343 2.147052 2.060151 1.7783644 1.745872
#> [2,] 0.8724749 0.8823482 0.9165253 1.045295 1.270363 0.6747801 0.700846
#>        [,268]   [,269]    [,270]   [,271]    [,272]    [,273]    [,274]
#> [1,] 2.239432 2.147103 2.1664742 2.202935 2.1353264 2.1198070 1.7943637
#> [2,] 1.153024 1.102157 0.9105379 1.295948 0.9143243 0.7723502 0.9238505
#>         [,275]   [,276]   [,277]    [,278]   [,279]    [,280]    [,281]  [,282]
#> [1,] 1.8493939 2.137634 2.145169 1.9459437 1.863399 1.9183223 1.9236169 2.12019
#> [2,] 0.8982228 1.083604 1.011439 0.7510046 1.049800 0.6721966 0.9753596 1.40808
#>         [,283]   [,284]   [,285]    [,286]    [,287]    [,288]    [,289]
#> [1,] 1.9582133 2.110038 1.910789 2.0486933 1.9228816 2.1423035 1.8052090
#> [2,] 0.9137323 1.271350 1.163263 0.8176447 0.8316738 0.9218289 0.6798731
#>        [,290]    [,291]    [,292]    [,293]    [,294]   [,295]   [,296]
#> [1,] 2.079607 1.9333020 1.9889645 2.0099931 1.8606685 2.102083 2.199250
#> [2,] 0.656936 0.8266694 0.9181029 0.9686097 0.7961075 1.057910 1.320519
#>        [,297]    [,298]   [,299]   [,300]    [,301]   [,302]   [,303]    [,304]
#> [1,] 2.327367 1.8113759 2.093617 2.068928 1.9903947 2.222476 2.066399 1.7182242
#> [2,] 1.245506 0.7201799 1.435428 1.169195 0.9748851 1.267682 1.001412 0.7987328
#>         [,305]   [,306]   [,307]   [,308]   [,309]   [,310]   [,311]   [,312]
#> [1,] 1.8193123 2.053442 1.968010 1.889622 2.125970 1.804124 2.244182 2.040007
#> [2,] 0.6008019 1.361036 0.786447 0.870576 1.108629 1.101373 1.370264 1.384107
#>        [,313]    [,314]    [,315]   [,316]    [,317]   [,318]   [,319]
#> [1,] 2.137316 1.8734479 1.9594064 2.145107 1.8244234 2.067851 2.187875
#> [2,] 1.097459 0.9132648 0.7991317 1.052694 0.7735229 0.627660 1.479166
#>         [,320]   [,321]   [,322]   [,323]   [,324]   [,325]    [,326]   [,327]
#> [1,] 1.8626148 2.021218 2.225771 2.039952 2.340267 1.999325 1.8174796 2.006622
#> [2,] 0.8574107 1.290692 1.309740 1.497808 1.001229 1.015047 0.8507137 1.008389
#>         [,328]    [,329]    [,330]   [,331]   [,332]   [,333]    [,334]
#> [1,] 2.0874505 2.0732823 1.9150033 2.137913 2.194852 1.993762 1.9064996
#> [2,] 0.9169987 0.8792006 0.8783308 1.241461 1.013770 1.097528 0.8005688
#>         [,335]   [,336]    [,337]   [,338]    [,339]   [,340]   [,341]   [,342]
#> [1,] 1.9112208 2.186797 1.9663679 2.126247 1.7900609 1.926468 1.951584 2.116295
#> [2,] 0.7505842 1.258050 0.8030775 1.242995 0.7139523 0.745247 1.057736 1.219186
#>        [,343]    [,344]   [,345]    [,346]   [,347]   [,348]    [,349]   [,350]
#> [1,] 2.092623 2.1293232 1.891918 1.8942994 2.088055 2.249072 1.9994702 2.122098
#> [2,] 1.182514 0.9088069 0.801880 0.8524203 1.148006 1.212627 0.8362871 1.118664
#>        [,351]   [,352]    [,353]    [,354]   [,355]   [,356]    [,357]   [,358]
#> [1,] 2.333112 2.062227 2.0435059 1.9711936 2.160584 2.107695 1.9309158 1.966092
#> [2,] 1.127517 1.239586 0.5844585 0.6354777 1.073000 1.119831 0.7329773 1.087604
#>        [,359]    [,360]    [,361]    [,362]    [,363]   [,364]    [,365]
#> [1,] 1.967607 2.0088490 1.9260543 1.8489241 2.3148764 2.264078 1.8716342
#> [2,] 1.278436 0.4924842 0.8721149 0.8166611 0.9848272 1.334832 0.8519732
#>        [,366]   [,367]   [,368]    [,369]    [,370]   [,371]    [,372]
#> [1,] 1.768645 1.802956 2.133193 2.0570290 2.0461603 2.182351 1.8971560
#> [2,] 0.920076 1.212440 1.138557 0.9536461 0.9394779 1.242223 0.8445213
#>         [,373]    [,374]    [,375]   [,376]    [,377]    [,378]    [,379]
#> [1,] 1.9524376 1.7163308 1.8416815 2.089898 2.0823704 2.1342934 1.9757384
#> [2,] 0.8342418 0.8543731 0.9105506 1.157171 0.7354355 0.8265668 0.8771057
#>         [,380]   [,381]    [,382]    [,383]   [,384]   [,385]   [,386]   [,387]
#> [1,] 2.2569382 2.094889 1.9394333 1.8695406 2.148128 2.150506 2.251177 2.081574
#> [2,] 0.9488837 1.418209 0.9558494 0.7965058 1.232952 1.045256 1.062280 1.298082
#>        [,388]   [,389]    [,390]   [,391]    [,392]    [,393]   [,394]   [,395]
#> [1,] 2.063309 1.981907 1.8896679 2.082698 1.8268750 1.8746162 2.259504 2.133818
#> [2,] 1.262899 1.224690 0.8844239 1.162865 0.8230154 0.9835858 1.190111 1.301929
#>        [,396]   [,397]   [,398]   [,399]   [,400]   [,401]    [,402]   [,403]
#> [1,] 1.975873 1.832048 1.969893 2.262028 2.020498 2.173219 2.0566595 2.106663
#> [2,] 1.062312 0.950685 1.128127 1.340725 1.011720 1.155384 0.6468319 1.011128
#>        [,404]    [,405]    [,406]    [,407]    [,408]   [,409]   [,410]
#> [1,] 1.727436 2.0298273 2.0640815 1.9065516 1.9069259 2.114642 2.111405
#> [2,] 0.847827 0.8747971 0.7913569 0.8475618 0.7246354 1.487401 1.075064
#>        [,411]    [,412]    [,413]    [,414]    [,415]   [,416]    [,417]
#> [1,] 2.144933 1.9207867 1.9592204 1.8910155 1.8586600 1.928694 1.9707356
#> [2,] 1.344222 0.9400245 0.5953241 0.7017105 0.7978047 0.663531 0.6042837
#>        [,418]    [,419]    [,420]   [,421]    [,422]   [,423]   [,424]
#> [1,] 2.328593 2.0670346 1.6140705 1.616348 2.0479700 2.017453 2.022570
#> [2,] 1.193573 0.8325329 0.8473604 0.765403 0.8103604 1.222377 0.901743
#>         [,425]    [,426]   [,427]    [,428]   [,429]   [,430]    [,431]
#> [1,] 1.9472204 1.9767715 1.984020 1.9933683 2.002989 1.751596 1.9881186
#> [2,] 0.6289196 0.7958221 0.700446 0.9087878 1.139704 1.012285 0.8532903
#>        [,432]    [,433]    [,434]   [,435]    [,436]    [,437]   [,438]  [,439]
#> [1,] 2.042433 1.9180783 2.3511903 2.084354 1.7862695 2.0628745 2.067445 1.98701
#> [2,] 0.929316 0.7660892 0.9140781 1.133850 0.8324375 0.9942276 1.168523 0.80416
#>        [,440]    [,441]    [,442]    [,443]    [,444]   [,445]   [,446]
#> [1,] 2.260974 2.0013447 1.9261268 2.1849801 1.7553970 1.988813 2.027242
#> [2,] 1.044172 0.9439994 0.7668938 0.8659774 0.6002489 1.160289 1.093194
#>        [,447]    [,448]    [,449]   [,450]    [,451]   [,452]    [,453]
#> [1,] 2.068849 2.0551605 1.9215158 2.139130 2.0613794 2.043308 2.0338085
#> [2,] 1.337046 0.9783186 0.6585334 1.167991 0.8861711 1.396965 0.9565224
#>        [,454]   [,455]   [,456]   [,457]    [,458]   [,459]   [,460]   [,461]
#> [1,] 2.200528 2.017695 1.706097 1.985281 1.8596301 2.198948 2.284493 2.119224
#> [2,] 1.105788 1.275051 0.846408 1.121783 0.9342835 1.141012 1.182233 1.207538
#>        [,462]   [,463]  [,464]    [,465]    [,466]   [,467]   [,468]    [,469]
#> [1,] 2.025658 1.895257 2.08153 1.9022285 1.7164376 1.862813 2.126600 1.9058931
#> [2,] 1.111444 0.849173 1.01027 0.6411341 0.5901954 0.821277 1.135292 0.6683009
#>        [,470]   [,471]    [,472]   [,473]   [,474]   [,475]   [,476]   [,477]
#> [1,] 1.924738 2.192261 1.7664244 2.134635 2.291793 2.001531 1.988303 2.176803
#> [2,] 1.179154 1.267134 0.9071426 1.242394 1.329121 1.135061 1.012029 1.222232
#>         [,478]    [,479]    [,480]   [,481]   [,482]    [,483]    [,484]
#> [1,] 1.6897799 1.7741883 2.2062686 2.101024 2.073307 2.0073971 1.8937816
#> [2,] 0.9536044 0.5489604 0.9833894 1.236851 1.179797 0.7843268 0.8789806
#>         [,485]   [,486]   [,487]    [,488]   [,489]   [,490]    [,491]
#> [1,] 2.1576664 2.097388 1.876600 1.7942653 2.132043 2.105732 2.0628508
#> [2,] 0.8179021 1.209710 1.159294 0.7664045 1.227551 1.191846 0.7733776
#>         [,492]    [,493]    [,494]    [,495]   [,496]   [,497]    [,498]
#> [1,] 1.9034088 1.8492070 1.9417222 2.0952427 2.142524 2.209316 1.8901491
#> [2,] 0.9338821 0.7805099 0.8082609 0.8635622 1.224887 1.161517 0.6960018
#>        [,499]    [,500]
#> [1,] 2.004936 2.0602100
#> [2,] 0.910550 0.6799487